+++ plutus-ledger-api/src/v1/address.rs
@@ -4,15 +4,20 @@ use std::str::FromStr;
 use anyhow::anyhow;
 use cardano_serialization_lib as csl;
 
+use crate::csl::csl_to_pla::{FromCSL, TryFromCSL, TryFromCSLError, TryToPLA};
+use crate::csl::pla_to_csl::{TryFromPLA, TryFromPLAError, TryToCSL};
 use crate::plutus_data::{
     verify_constr_fields, IsPlutusData, PlutusData, PlutusDataError, PlutusType,
 };
-use crate::utils::csl_to_pla::TryToPLA;
 use crate::v1::crypto::Ed25519PubKeyHash;
 use crate::v1::script::ValidatorHash;
 #[cfg(feature = "lbf")]
 use lbr_prelude::json::{self, Error, Json};
 use num_bigint::BigInt;
+
+/////////////
+// Address //
+/////////////
 
 #[cfg(feature = "serde")]
 use serde::{Deserialize, Serialize};
@@ -28,6 +33,15 @@ use serde::{Deserialize, Serialize};
 pub struct Address {
     pub credential: Credential,
     pub staking_credential: Option<StakingCredential>,
+}
+
+impl Address {
+    pub fn with_extra_info(&self, network_tag: u8) -> AddressWithExtraInfo {
+        AddressWithExtraInfo {
+            address: self,
+            network_tag,
+        }
+    }
 }
 
 impl IsPlutusData for Address {
@@ -78,6 +92,75 @@ impl FromStr for Address {
             .map_err(|err| anyhow!("Couldn't convert address: {}", err))
     }
 }
+
+impl TryFromCSL<csl::address::Address> for Address {
+    fn try_from_csl(value: &csl::address::Address) -> Result<Self, TryFromCSLError> {
+        if let Some(addr) = csl::address::BaseAddress::from_address(value) {
+            Ok(Address {
+                credential: Credential::from_csl(&addr.payment_cred()),
+                staking_credential: Some(StakingCredential::from_csl(&addr.stake_cred())),
+            })
+        } else if let Some(addr) = csl::address::PointerAddress::from_address(value) {
+            Ok(Address {
+                credential: Credential::from_csl(&addr.payment_cred()),
+                staking_credential: Some(StakingCredential::from_csl(&addr.stake_pointer())),
+            })
+        } else if let Some(addr) = csl::address::EnterpriseAddress::from_address(value) {
+            Ok(Address {
+                credential: Credential::from_csl(&addr.payment_cred()),
+                staking_credential: None,
+            })
+        } else {
+            Err(TryFromCSLError::ImpossibleConversion(format!(
+                "Unable to represent address {:?}",
+                value
+            )))
+        }
+    }
+}
+
+#[derive(Clone, Debug)]
+pub struct AddressWithExtraInfo<'a> {
+    pub address: &'a Address,
+    pub network_tag: u8,
+}
+
+impl TryFromPLA<AddressWithExtraInfo<'_>> for csl::address::Address {
+    fn try_from_pla(val: &AddressWithExtraInfo<'_>) -> Result<Self, TryFromPLAError> {
+        let payment = val.address.credential.try_to_csl()?;
+
+        Ok(match val.address.staking_credential {
+            None => csl::address::EnterpriseAddress::new(val.network_tag, &payment).to_address(),
+            Some(ref sc) => match sc {
+                StakingCredential::Hash(c) => {
+                    csl::address::BaseAddress::new(val.network_tag, &payment, &c.try_to_csl()?)
+                        .to_address()
+                }
+                StakingCredential::Pointer(ptr) => {
+                    csl::address::PointerAddress::new(val.network_tag, &payment, &ptr.try_to_csl()?)
+                        .to_address()
+                }
+            },
+        })
+    }
+}
+
+impl std::fmt::Display for AddressWithExtraInfo<'_> {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        let bech32_addr: Option<String> = self
+            .try_to_csl()
+            .ok()
+            .and_then(|csl_addr: csl::address::Address| csl_addr.to_bech32(None).ok());
+        match bech32_addr {
+            Some(addr) => write!(f, "{}", addr),
+            None => write!(f, "INVALID ADDRESS {:?}", self),
+        }
+    }
+}
+
+////////////////
+// Credential //
+////////////////
 
 /// A public key hash or validator hash credential (used as a payment or a staking credential)
 #[derive(Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
@@ -172,6 +255,36 @@ impl Json for Credential {
         )
     }
 }
+
+impl FromCSL<csl::address::StakeCredential> for Credential {
+    fn from_csl(value: &csl::address::StakeCredential) -> Self {
+        match value.kind() {
+            csl::address::StakeCredKind::Key => {
+                Credential::PubKey(Ed25519PubKeyHash::from_csl(&value.to_keyhash().unwrap()))
+            }
+            csl::address::StakeCredKind::Script => {
+                Credential::Script(ValidatorHash::from_csl(&value.to_scripthash().unwrap()))
+            }
+        }
+    }
+}
+
+impl TryFromPLA<Credential> for csl::address::StakeCredential {
+    fn try_from_pla(val: &Credential) -> Result<Self, TryFromPLAError> {
+        match val {
+            Credential::PubKey(pkh) => Ok(csl::address::StakeCredential::from_keyhash(
+                &pkh.try_to_csl()?,
+            )),
+            Credential::Script(sh) => Ok(csl::address::StakeCredential::from_scripthash(
+                &sh.0.try_to_csl()?,
+            )),
+        }
+    }
+}
+
+///////////////////////
+// StakingCredential //
+///////////////////////
 
 /// Credential (public key hash or pointer) used for staking
 #[derive(Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
@@ -278,6 +391,48 @@ impl Json for StakingCredential {
     }
 }
 
+impl FromCSL<csl::address::StakeCredential> for StakingCredential {
+    fn from_csl(value: &csl::address::StakeCredential) -> Self {
+        StakingCredential::Hash(Credential::from_csl(value))
+    }
+}
+
+impl TryFromPLA<StakingCredential> for csl::address::StakeCredential {
+    fn try_from_pla(val: &StakingCredential) -> Result<Self, TryFromPLAError> {
+        match val {
+            StakingCredential::Hash(c) => c.try_to_csl(),
+            StakingCredential::Pointer(_) => Err(TryFromPLAError::ImpossibleConversion(
+                "cannot represent chain pointer".into(),
+            )),
+        }
+    }
+}
+
+impl FromCSL<csl::address::Pointer> for StakingCredential {
+    fn from_csl(value: &csl::address::Pointer) -> Self {
+        StakingCredential::Pointer(ChainPointer::from_csl(value))
+    }
+}
+
+#[derive(Clone, Debug)]
+pub struct RewardAddressWithExtraInfo<'a> {
+    pub staking_credential: &'a StakingCredential,
+    pub network_tag: u8,
+}
+
+impl TryFromPLA<RewardAddressWithExtraInfo<'_>> for csl::address::RewardAddress {
+    fn try_from_pla(val: &RewardAddressWithExtraInfo<'_>) -> Result<Self, TryFromPLAError> {
+        Ok(csl::address::RewardAddress::new(
+            val.network_tag,
+            &val.staking_credential.try_to_csl()?,
+        ))
+    }
+}
+
+//////////////////
+// ChainPointer //
+//////////////////
+
 /// In an address, a chain pointer refers to a point of the chain containing a stake key
 /// registration certificate. A point is identified by 3 coordinates:
 /// - An absolute slot number
@@ -291,6 +446,30 @@ pub struct ChainPointer {
     pub transaction_index: TransactionIndex,
     pub certificate_index: CertificateIndex,
 }
+
+impl FromCSL<csl::address::Pointer> for ChainPointer {
+    fn from_csl(value: &csl::address::Pointer) -> Self {
+        ChainPointer {
+            slot_number: Slot::from_csl(&value.slot_bignum()),
+            transaction_index: TransactionIndex::from_csl(&value.tx_index_bignum()),
+            certificate_index: CertificateIndex::from_csl(&value.cert_index_bignum()),
+        }
+    }
+}
+
+impl TryFromPLA<ChainPointer> for csl::address::Pointer {
+    fn try_from_pla(val: &ChainPointer) -> Result<Self, TryFromPLAError> {
+        Ok(csl::address::Pointer::new_pointer(
+            &val.slot_number.try_to_csl()?,
+            &val.transaction_index.try_to_csl()?,
+            &val.certificate_index.try_to_csl()?,
+        ))
+    }
+}
+
+//////////
+// Slot //
+//////////
 
 /// Number of slots elapsed since genesis
 #[derive(Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
@@ -308,6 +487,22 @@ impl IsPlutusData for Slot {
     }
 }
 
+impl FromCSL<csl::utils::BigNum> for Slot {
+    fn from_csl(value: &csl::utils::BigNum) -> Self {
+        Slot(BigInt::from_csl(value))
+    }
+}
+
+impl TryFromPLA<Slot> for csl::utils::BigNum {
+    fn try_from_pla(val: &Slot) -> Result<Self, TryFromPLAError> {
+        val.0.try_to_csl()
+    }
+}
+
+//////////////////////
+// CertificateIndex //
+//////////////////////
+
 /// Position of the certificate in a certain transaction
 #[derive(Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
 #[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
@@ -324,6 +519,22 @@ impl IsPlutusData for CertificateIndex {
     }
 }
 
+impl FromCSL<csl::utils::BigNum> for CertificateIndex {
+    fn from_csl(value: &csl::utils::BigNum) -> Self {
+        CertificateIndex(BigInt::from_csl(value))
+    }
+}
+
+impl TryFromPLA<CertificateIndex> for csl::utils::BigNum {
+    fn try_from_pla(val: &CertificateIndex) -> Result<Self, TryFromPLAError> {
+        val.0.try_to_csl()
+    }
+}
+
+//////////////////////
+// TransactionIndex //
+//////////////////////
+
 /// Position of a transaction in a given slot
 /// This is not identical to the index of a `TransactionInput`
 #[derive(Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
@@ -338,5 +549,17 @@ impl IsPlutusData for TransactionIndex {
 
     fn from_plutus_data(data: &PlutusData) -> Result<Self, PlutusDataError> {
         IsPlutusData::from_plutus_data(data).map(Self)
+    }
+}
+
+impl FromCSL<csl::utils::BigNum> for TransactionIndex {
+    fn from_csl(value: &csl::utils::BigNum) -> Self {
+        TransactionIndex(BigInt::from_csl(value))
+    }
+}
+
+impl TryFromPLA<TransactionIndex> for csl::utils::BigNum {
+    fn try_from_pla(val: &TransactionIndex) -> Result<Self, TryFromPLAError> {
+        val.0.try_to_csl()
     }
 }
