+++ plutus-ledger-api/src/v1/assoc_map.rs
@@ -8,6 +8,10 @@ use serde::{Deserialize, Serialize};
 
 use crate::plutus_data::{IsPlutusData, PlutusData, PlutusDataError, PlutusType};
 
+//////////////
+// AssocMap //
+//////////////
+
 #[derive(Debug, PartialEq, Eq, Clone)]
 #[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
 pub struct AssocMap<K, V>(pub Vec<(K, V)>);
