+++ plutus-ledger-api/src/v1/datum.rs
@@ -1,4 +1,9 @@
 //! Types related to Plutus Datums
+
+use cardano_serialization_lib as csl;
+
+use crate::csl::csl_to_pla::FromCSL;
+use crate::csl::pla_to_csl::{TryFromPLA, TryFromPLAError, TryToCSL};
 use crate::plutus_data::{IsPlutusData, PlutusData, PlutusDataError};
 use crate::v1::crypto::LedgerBytes;
 #[cfg(feature = "lbf")]
@@ -6,6 +11,10 @@ use lbr_prelude::json::Json;
 
 #[cfg(feature = "serde")]
 use serde::{Deserialize, Serialize};
+
+///////////////
+// DatumHash //
+///////////////
 
 /// blake2b-256 hash of a datum
 #[derive(Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
@@ -23,6 +32,23 @@ impl IsPlutusData for DatumHash {
     }
 }
 
+impl FromCSL<csl::crypto::DataHash> for DatumHash {
+    fn from_csl(value: &csl::crypto::DataHash) -> Self {
+        DatumHash(LedgerBytes(value.to_bytes()))
+    }
+}
+
+impl TryFromPLA<DatumHash> for csl::crypto::DataHash {
+    fn try_from_pla(val: &DatumHash) -> Result<Self, TryFromPLAError> {
+        csl::crypto::DataHash::from_bytes(val.0 .0.to_owned())
+            .map_err(TryFromPLAError::CSLDeserializeError)
+    }
+}
+
+///////////
+// Datum //
+///////////
+
 /// Piece of information associated with a UTxO encoded into a PlutusData type.
 #[derive(Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
 #[cfg_attr(feature = "lbf", derive(Json))]
@@ -36,5 +62,11 @@ impl IsPlutusData for Datum {
 
     fn from_plutus_data(data: &PlutusData) -> Result<Self, PlutusDataError> {
         IsPlutusData::from_plutus_data(data).map(Self)
+    }
+}
+
+impl TryFromPLA<Datum> for csl::plutus::PlutusData {
+    fn try_from_pla(val: &Datum) -> Result<Self, TryFromPLAError> {
+        val.0.try_to_csl()
     }
 }
